Abins refactoring: downcasing, replace factory with pure function
- Continue downcasing module names for easier distinction from classes

- InstrumentProducer is just a namespace for a staticmethod; we can
  use a module for that. Replace with a pure function
  abins.instruments.get_instrument(name).

  - There are a couple of compromises being made here:

    - We don't always like to put important things into __init__ files
      because it makes them harder to find. But it makes sense for this to
      sit with the instruments rather than at top-level, and it's job
      is ultimately to fetch things from abins.instruments which is
      something __init__ imports also do.

    - The name get_instrument() is slightly awkward as this is not
      exactly a "getter"; perhaps produce_instrument() would indeed be
      better. But it _is_ intuitive that get_instrument() will
      return an instantiated Instrument object, and this is the
      important thing to communicate.
