+++ scripts/abins/spowdersemiempiricalcalculator.py
